+++ scripts/test.js
@@ -35,6 +35,15 @@ if(process.platform == 'darwin')
 	// 7zip
 	exists('./node_modules/7zip-bin-full/mac/arm64/7zz', fs.constants.X_OK | fs.constants.R_OK);
 	exists('./node_modules/7zip-bin-full/mac/x64/7zz', fs.constants.X_OK | fs.constants.R_OK);
+
+	// OpenComicAI
+	exists('./node_modules/opencomic-ai-bin/mac/arm64/realcugan/realcugan-ncnn-vulkan.app', fs.constants.X_OK | fs.constants.R_OK);
+	exists('./node_modules/opencomic-ai-bin/mac/arm64/waifu2x/waifu2x-ncnn-vulkan.app', fs.constants.X_OK | fs.constants.R_OK);
+	exists('./node_modules/opencomic-ai-bin/mac/arm64/upscayl/upscayl-bin.app', fs.constants.X_OK | fs.constants.R_OK);
+	exists('./node_modules/opencomic-ai-bin/mac/x64/realcugan/realcugan-ncnn-vulkan.app', fs.constants.X_OK | fs.constants.R_OK);
+	exists('./node_modules/opencomic-ai-bin/mac/x64/waifu2x/waifu2x-ncnn-vulkan.app', fs.constants.X_OK | fs.constants.R_OK);
+	exists('./node_modules/opencomic-ai-bin/mac/x64/upscayl/upscayl-bin.app', fs.constants.X_OK | fs.constants.R_OK);
+
 }
 else if(process.platform == 'linux')
 {	
@@ -55,6 +64,13 @@ else if(process.platform == 'linux')
 	exists('./node_modules/7zip-bin-full/linux/arm64/7zz', fs.constants.X_OK | fs.constants.R_OK);
 	exists('./node_modules/7zip-bin-full/linux/ia32/7zz', fs.constants.X_OK | fs.constants.R_OK);
 	exists('./node_modules/7zip-bin-full/linux/x64/7zz', fs.constants.X_OK | fs.constants.R_OK);
+
+	// OpenComicAI
+	exists('./node_modules/opencomic-ai-bin/linux/arm64/realcugan/realcugan-ncnn-vulkan', fs.constants.X_OK | fs.constants.R_OK);
+	exists('./node_modules/opencomic-ai-bin/linux/arm64/waifu2x/waifu2x-ncnn-vulkan', fs.constants.X_OK | fs.constants.R_OK);
+	exists('./node_modules/opencomic-ai-bin/linux/x64/realcugan/realcugan-ncnn-vulkan', fs.constants.X_OK | fs.constants.R_OK);
+	exists('./node_modules/opencomic-ai-bin/linux/x64/waifu2x/waifu2x-ncnn-vulkan', fs.constants.X_OK | fs.constants.R_OK);
+	exists('./node_modules/opencomic-ai-bin/linux/x64/upscayl/upscayl-bin', fs.constants.X_OK | fs.constants.R_OK);
 }
 else if(process.platform == 'win32')
 {
@@ -76,6 +92,11 @@ else if(process.platform == 'win32')
 	exists('./node_modules/7zip-bin-full/win/x64/7z.dll', fs.constants.X_OK | fs.constants.R_OK);
 	exists('./node_modules/7zip-bin-full/win/arm64/7z.exe', fs.constants.X_OK | fs.constants.R_OK);
 	exists('./node_modules/7zip-bin-full/win/arm64/7z.dll', fs.constants.X_OK | fs.constants.R_OK);
+
+	// OpenComicAI
+	exists('./node_modules/opencomic-ai-bin/win/x64/realcugan/realcugan-ncnn-vulkan.exe', fs.constants.X_OK | fs.constants.R_OK);
+	exists('./node_modules/opencomic-ai-bin/win/x64/waifu2x/waifu2x-ncnn-vulkan.exe', fs.constants.X_OK | fs.constants.R_OK);
+	exists('./node_modules/opencomic-ai-bin/win/x64/upscayl/upscayl-bin.exe', fs.constants.X_OK | fs.constants.R_OK);
 }
 
 console.log('Runed tests: Ok');